Task 4 code: Student/exam association
I finished the 4th task. I made a simple command prompt interactive application that assigns students to exams.

For aspectJ and reuse, I just copy pasted the Loggable.aj and Connector.aj. But I had to change the connector with my own class names. It would be nice if I didn't need to touch the aspect. That is especially the case if I had 30 classes that would need logging. ( && (within(class1)) && (within(class2)) && ... ) is long to write. An "implements" on the needed classes is more elegant. :)

It is a coincidence that I have 2. That's why (-> Elvric) I suggest creating an interface that has literally nothing in it and change the aspect to act on all implementations of that interface, say "Logged.java". As for the Logger.log() function, I created my own. But I think that we should make the compiler produce an error if that class and that function are not declared.

In conclusion, the code and aspect reuse worked. But there are some possible improvements.